requirements.txt
tensorflow==2.15.0
fastapi
uvicorn
pillow
numpy
mlflow
